requirements.txt
dotenv
librosa
assemblyai

json
datetime
google.generativeai
os
faiss-cpu
tqdm
numpy
transformers
uuid